% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_epp_wpp}
\alias{age_epp_wpp}
\title{Within pair paternity data and age}
\format{A raw csv file containing the following columns and descriptors.}
\description{
Proportion of within and extra-pair paternity for different ages. Data taken from Cleasby and Nakagawa (2012).
}
\references{
Cleasby, I.R. and Nakagawa, S. (2012) The influence of male age on within-pair and extra-pair paternity in passerines. Ibis, 154: 318-324.
}

