% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birdegg_ssd.8}
\alias{birdegg_ssd.8}
\title{birdegg_ssd.8 data}
\format{
A raw csv file containing the following columns and descriptors
\enumerate{
\item animal:	species Latin name
\item Spp:	species common name
\item Lndim:	logaritm of adult mass dimorphism
\item Measure:	method of eggs measurments
\item Neggs:	number of eggs
\item Nclutches:	number of clutches
\item ESr:	effect size, correlation
\item Type:	how the effect size was calculated: row data/inferential statistics
\item StudyID:	reference
\item Year:	year of publication
\item D:	design effect
\item EN:	effective sample size
\item Zr:	Fisher's Z
\item VZr:	sampling variance
}
}
\description{
Data set on bird egg sexual size dimorphism meta-analysis. Data was taken from Rutkowska et al. (2014). In this study more than one egg came from the same study which introduces non-independence in the data (i.e. correlated structure). As a sensitivity analysis Rutkowska et al. (2014) used an effective sample size proposed by Higgins and Green (2009), which requires an estimate of the intra-class correlation coefficient (ICC) in the calculation of effect sizes. In this case, an ICC estimate of 0.8 was used.
}
\references{
Rutkowska, J., Dubiec, A. and Nakgawa, S. (2014). All eggs are made equal: a meta-analysis of egg sexual size dimorphism. Journal of Evolutionary Biology, 27:153-160

Higgins, J.P.T. and Green, S. (2009). Cochran handbook for systematic reviews and interventions. Wiley-Blackwell. pgs 412-414.
}
