% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMatrix.R
\name{cov_matrix}
\alias{cov_matrix}
\title{Covariance and correlation matrix function}
\usage{
cov_matrix(data, V, var1, var2 = FALSE, cor, type = c("sc", "ws"))
}
\arguments{
\item{data}{The data object that a correlation or covariance matrix needs to be generated for.}

\item{V}{Character strong indicating the column name of the known sampling error variance.}

\item{var1}{The vector describing the first variable each effect size is derived from. For example, this could be study ID.}

\item{var2}{An optional second variable that will further split the data. For example, by trait type. If var2 is not provided it is assumed only var1 is of interest.}

\item{cor}{The known or hypothesized correlation between effect sizes resulting from within-study non-independence. cor values could be derived from raw data or from known correlations taken from the literature. Alternatively, multiple cor values can be used to derive several matrices each of which can be used in a sensitivity analysis.}

\item{type}{Whether a general within study blocked matrix is needed ("ws" - note this could be based on more than one variable), or a shared control covariance matrix is needed ("sc").}
}
\description{
Function for generating various covariance and correlation matrices used to control for sources of non-independence in meta-analyses. Currently only shared-control and general forms of within-study covariances (e.g. shared traits - for sensitivity analysis) can be conducted. More complex matrices are not yet implemented.
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}

