% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coMatrix.R
\name{covMatrix}
\alias{covMatrix}
\title{Generate a covariance matrix}
\usage{
covMatrix(data, es_var, cor)
}
\arguments{
\item{data}{The data object that a correlation or covariance matrix needs to be generated for.}

\item{es_var}{Character string indicating the column name of the known sampling error variance.}

\item{cor}{The known or hypothesized correlation between effect sizes resulting from within-study non-independence. cor values could be derived from raw data or from known correlations taken from the literature. Alternatively, multiple cor values can be used to derive several matrices each of which can be used in a sensitivity analysis.}
}
\description{
This is a sub-function used by cov_matrix which generates a covariance matrix for a sub-set of a data frame containing the effect size sampling variance and a known correlation matrix
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}

