% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heatshock}
\alias{heatshock}
\title{heat shock and life extension data}
\format{A raw csv file containing the following columns and descriptors}
\description{
Data set investing the impact of heat shock exposure on longevity. Data was taken from Lagisz et al. (2013).
}
\references{
Lagisz, M., Hector, K.L. and Nakagawa, S. (2013) Life exstension after heat shock exposure: Assessing meta-analytic evidence for hormesis. Ageing Research Reviews, 12:653-660.
}

