% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{hedge}
\alias{hedge}
\title{hedge}
\usage{
hedge(m1, m2, sd1, sd2, n1, n2, type = c("d", "g"))
}
\arguments{
\item{m1}{Mean of group 1}

\item{m2}{Mean of group 2}

\item{sd1}{Standard deviation of group 1}

\item{sd2}{Standard deviation of group 2}

\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}

\item{type}{Whether Hedges' d (type = "d") and biased corrected Hedges' d or Hedges' g (type = "g") is the be used.}
}
\description{
Function for calculating Hedges' d or biased corrected Hedges' g. m1 is subtracted from m2.
}
\references{
Borenstein et al. 2009. Effect sizes based on means. In: Introduction to Meta-Analysis (eds. Borenstein, M., Hedges, L.V., Higgins, J.P.T. and Rothstein, H.R.). pg 21-32.
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}
