% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2calc.R
\name{I2}
\alias{I2}
\title{I^2 Function}
\usage{
I2(model, v, re.list = list(phylo = "animal", spp = "species", stdy =
  "study"))
}
\arguments{
\item{model}{The MCMCglmm model object}

\item{v}{The vector of sampling variance for each effect size.}

\item{re.list}{A list of variance estimates one wishes to derive I^2 estimates for. These could include phylogenetic heritability (H^2), species ($I^{2}_{sp}$) and study ($I^{2}_{stdy}$) or often called ($\tau^{2}$). At the moment only three types are provided: phylogenetic (phylo), species (spp) and study (stdy). The names of your specific variance component do not matter, but these names should be specified in the re.list argument in the respective argument.}
}
\value{
A data.frame containing the relevant I^2 measures along with the 95 percent credible intervals for each element listed in the re.list argument
}
\description{
Function for calculating commonly reported  I^2 measures from MCMCglmm model objects.  See Nakagawa and Santos (2012) for detailed explanation on the various types of I^2
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}
\references{
Nakagawa, S. and Santos, E.S.A. (2012) Methodological issues and advances in biological meta-analysis. Evolutionary Ecology, 26:1253-1274.
}

