% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2calc.R
\name{I2calc}
\alias{I2calc}
\title{$I^2$ Function}
\usage{
I2calc(model, v, group = FALSE, re.list = list(phylo = "phylo", spp = "spp",
  stdy = "Paper_No"))
}
\arguments{
\item{model}{The MCMCglmm model object}

\item{v}{The vector of sampling variance for each effect size.}

\item{group}{The specific variance components one wishes to extract. Note that mev might be included in the model or as part of ginverse and these will be excluded if group = FALSE}

\item{re.list}{A list of variance estimates one wishes to derive $I^2$ estimates for. These could include phylogenetic heritability ($H^2$), species ($I^2_{sp}$) and study (\eqn{I^2_{stdy}} or often called \eqn{tau^2}).}
}
\value{
A data.frame containing the relevant I^2 measures along with the 95$%$ credible intervals for each element listed in the re.list argument
}
\description{
Function for calculating \eqn{I^2} from MCMCglmm model objects according to Nakagawa and Santos (2012)
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}
\references{
Nakagawa, S. and Santos, E.S.A. (2012) Methodological issues and advances in biological meta-analysis. Evolutionary Ecology, 26:1253-1274.
}

