% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{lnCVR_es}
\alias{lnCVR_es}
\title{lnCVR}
\usage{
lnCVR_es(m1, m2, sd1, sd2, n1, n2)
}
\arguments{
\item{m1}{Mean of group 1}

\item{m2}{Mean of group 2}

\item{sd1}{Standard deviation of group 1}

\item{sd2}{Standard deviation of group 2}

\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}
}
\description{
Function for calculating log coefficient of variation ratio
}
\author{
Alistair Senior - Alistair.senior@sydney.edu.au
}
\references{
Nakagawa et al. 2015.
}

