% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{lnHR_es}
\alias{lnHR_es}
\title{lnHR_es}
\usage{
lnHR_es(d1, d2, n1HR, n2HR)
}
\arguments{
\item{d1}{Number of events (e.g. deaths) in time interval t-1 to t for group 1}

\item{d2}{Number of events (e.g. deaths) in time interval t-1 to t for group 2}

\item{n1HR}{Number at risk for group 1 during time interval t-1 to t}

\item{n2HR}{Number at risk for group 2 during time interval t-1 to t}
}
\description{
Function for calculating log hazards ratio effect size statistics, for a specified period in time-to-event or survival curve data.
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}
\references{
Williamson PR, Smith CT, Hutton JL,  Marson AG (2002). Aggregate data meta-analysis with time-to-event outcomes. Stat. Med. 21, 3337-3351.

Parmar MKB, Torri V,  Stewart L (1998). Extracting summary statistics to perform meta-analyses of the published literature for survival endpoints. Stat. Med. 17, 2815-2834.
}

