% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{lnRR_es}
\alias{lnRR_es}
\title{lnRR}
\usage{
lnRR_es(m1, m2, sd1, sd2, n1, n2)
}
\arguments{
\item{m1}{Mean of group 1}

\item{m2}{Mean of group 2}

\item{sd1}{Standard deviation of group 1}

\item{sd2}{Standard deviation of group 2}

\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}
}
\description{
Function for calculating log response ratio
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}
