% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{lnSD_es}
\alias{lnSD_es}
\title{lnSD_es}
\usage{
lnSD_es(sd, n)
}
\arguments{
\item{sd}{Standard deviation of treatment}

\item{n}{Sample size of treatment}
}
\description{
Function for calculating log standard deviation effect size statistics
}
\author{
Alistair Senior - Alistair.senior@sydney.edu.au
}
\references{
Nakagawa et al. 2015.
}

