% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{lOR_es}
\alias{lOR_es}
\title{lnOR}
\usage{
lOR_es(p1, p2, n1, n2)
}
\arguments{
\item{p1}{Proportion in group 1}

\item{p2}{Proportion in group 2}

\item{n1}{Sample size in group 1}

\item{n2}{Sample size in group 2}
}
\description{
Function for calculating log odds ratio effect size statistics
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}

