% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_VCV_matrix.R
\name{make_VCV_matrix}
\alias{make_VCV_matrix}
\title{Covariance and correlation matrix function basing on shared level ID}
\usage{
make_VCV_matrix(data, V, cluster, obs, type = c("vcv", "cor"), rho = 0.5)
}
\arguments{
\item{data}{Dataframe object containing effect sizes, their variance, unique IDs and clustering variable}

\item{V}{Name of the variable (vector) containing effect size variances variances}

\item{cluster}{Name of the variable (vector) indicating which effects belong to the same cluster. Same value of 'cluster' are assumed to be nonindependent (correlated).}

\item{obs}{Name of the variable (vector) containing individual IDs for each value in the V (Vector of variances). If this parameter is missing, label will be labelled with consecutive integers starting from 1.}

\item{type}{Optional logical parameter indicating whether a full variance-covariance matrix (default or "vcv") is needed or a correlation matrix ("cor") for the non-independent blocks of variance values.}

\item{rho}{Known or assumed correlation value among effect sizes sharing same 'cluster' value. Default value is 0.5.}
}
\description{
Function for generating simple covariance and correlation matrices
}
