% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matdiet}
\alias{matdiet}
\title{matdiet}
\format{A raw csv file containing the following columns and descriptors}
\description{
Data set on how maternal diet impacts copying styles in rodents. Data was taken from Besson et al. (2016).
}
\references{
Besson, A. A., Lagisz, M., Senior, A.M., Hector, K.L. and Nakagawa, S. (2016). Effect of maternal diet on offspring coping styles in rodents: a systematic review and meta-analysis. Biological Reviews, 91:1065-1080
}
