% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moNumtradeoff}
\alias{moNumtradeoff}
\title{moNumtradeoff}
\format{A raw csv file containing the following columns and descriptors}
\description{
Data set investing the correlations between maternal size and offspring number (dataset 2 - M-O.Number). Data was taken from Lim et al. (2014).
}
\references{
Lim, J. N., Senior, A. M. & Nakagawa, S. (2014) Heterogeneity in individual quality and reproductive trade-offs within species. Evolution. 68: 2306–2318
}

