% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moSizetradeoff}
\alias{moSizetradeoff}
\title{moSizetradeoff}
\format{
A raw csv file containing the following columns and descriptors
}
\description{
Data set investing the correlations between maternal size and offspring size (dataset 1: MO.size). Data was taken from Lim et al. (2014).
}
\references{
Lim, J. N., Senior, A. M. and Nakagawa, S. (2014) Heterogeneity in individual quality and reproductive trade-offs within species. Evolution, 68:2306-2318.
}
