% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtor}
\alias{mtor}
\title{mtor data}
\format{A raw csv file containing the following columns and descriptors}
\description{
Data set investing whether changes in mTOR (rapamycin) signaling extend lifespan. Data was taken from Garratt et al. (2016).
}
\references{
Garratt, M., Nakagawa, S., Simons, M.J.P. (2016) Comparative idiosyncrasies in life extension by reduced mTOR signaling and its distinctiveness from dietary restriction. Aging Cell, 15: 737-743.
}
