% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ooAdjtradeoff}
\alias{ooAdjtradeoff}
\title{ooAdjtradeoff}
\format{A raw csv file containing the following columns and descriptors
\enumerate{
\item 
\item 
}}
\description{
Data set investing the correlations between offspring size and offspring number after adjusting for maternal size(dataset 3 - O-O.Adjusted). Data was taken from Lim et al. (2014).
}
\references{
Lim, J. N., Senior, A. M. & Nakagawa, S. (2014) Heterogeneity in individual quality and reproductive trade-offs within species. Evolution. 68: 2306–2318
}

