% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parasites}
\alias{parasites}
\title{Host manipulation and parasite data}
\format{
A raw csv file containing the following columns and descriptors. UI = Uninfected; U = Infected	
\enumerate{
\item Parasite.species: Species of parasite
\item Phylum: Phylum
\item Class: Class
\item Family: Family
}
}
\description{
Correlation between
}
\references{
Poulin, R. (2000) Manipulation of host behaviour by parasites: a weakening paradigm? Proceedings of the Royal Society B: Biological Sciences, 267: 1471-2954.
}
