% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plumage}
\alias{plumage}
\title{plumage and dominance}
\format{
A raw csv file containing the following columns and descriptors.
}
\description{
A meta-analysis of the relationship between plumage traits and dominance in birds. Data taken from Santos et al. (2011).
}
\references{
Santos, E.S.A., Scheck, D. and Nakagawa, S. (2011) Dominance and plumage traits: meta-analysis and meta-regression analysis. Animal Behaviour, 82: 3-19.
}
