% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2.R
\name{simMonteCarlo}
\alias{simMonteCarlo}
\title{Parametric simulation}
\usage{
simMonteCarlo(estimate, n, sims)
}
\arguments{
\item{estimate}{The estimate (i.e. variance) from a metafor model}

\item{n}{The sample size used in estimating the variance}

\item{sims}{The number of simulations}
}
\description{
Function for calculating I2 estimates using parametric simulations of model estimates taken from metafor. Note that the effectiveness of these simulations depends on the accuracy of model variance estimates.
}
\author{
Daniel Noble - daniel.noble@anu.edu.au
}
