% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2.R
\name{simulate}
\alias{simulate}
\title{Parametric simulation}
\usage{
simulate(estimate, sims)
}
\arguments{
\item{estimate}{The estimate (i.e. variance) from a metafor model}

\item{sims}{The number of simulations}
}
\description{
Function for calculating I2 estimates using parametric simulations of model estimates taken from metafor. Note that the effectiveness of these simulations depends on the accuracy of model variance estimates.
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}

