% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sparrows}
\alias{sparrows}
\title{sparrow data}
\format{A raw csv file containing the following columns and descriptors	
\enumerate{
\item StudyID: Study number
\item Place: Location of sparrow population
\item Correlation: correlation coefficient
\item SampleSize: Sample size for study
}}
\description{
Correlation between
}
\references{
Nakgawa, S., Ockendon, N., Gillespie, D.O.S., Hatchwell, B.J. and Burke, T. (2007) Assessing the function of house sparrows' bib size using a flexible meta-analysis method. Behavioural Ecology, 18:831-840.
}
