% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{var_hedge}
\alias{var_hedge}
\title{var_hedge}
\usage{
var_hedge(hedge, n1, n2, type = c("d", "g"))
}
\arguments{
\item{hedge}{Calculation of Hedges g or d}

\item{n1}{Sample size of treatment 1}

\item{n2}{Sample size of treatment 2}

\item{type}{Whether Hedges' d (type = "d") and biased corrected Hedges' d or Hedges' g (type = "g") is the be used.}
}
\description{
Function for calculating the sampling variance for Hedges' d or biased corrected Hedges' g.
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}

