% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{Zr_es}
\alias{Zr_es}
\title{Zr_es}
\usage{
Zr_es(r, n)
}
\arguments{
\item{r}{Correlation coefficient}

\item{n}{Sample size}
}
\description{
Function for calculating Fisher's z-transformed correlation.
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}

