% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{es_stat}
\alias{es_stat}
\title{es_stat}
\usage{
es_stat(m1, m2, sd1, sd2, n1, n2, p1, p2, r, nr, type = c("d", "g", "Zr",
  "lnOR", "lnRR", "lnCVR", "lnVR", "lnHR"))
}
\arguments{
\item{m1}{Mean of treatment 1}

\item{m2}{Mean of treatment 2}

\item{sd1}{Standard deviation of treatment 1}

\item{sd2}{Standard deviation of treatment 2}

\item{n1}{Sample size of treatment 1}

\item{n2}{Sample size of treatment 2}

\item{p1}{Proportion in treatment 1}

\item{p2}{Proportion in treatment 2}

\item{r}{Correlation coefficient}

\item{nr}{Sample size used for estimating the correlation coefficient}

\item{type}{The type specifies the specific effect statistics one wishes to calculate. Types include: "d", "g", "Zr","lnOR", "lnHR", "lnRR", "lnCVR", "lnVR".}
}
\description{
Function for calculating common effect size statistics.  Effect size statistics include both the effect size itself and it's sampling error. This function estimates both commonly used effect sizes (Hedges' d & g, Zr, lnOR, lnRR) along with less commonly used (lnHR) or newly developed effect sizes (i.e. for variance: lnCVR, lnVR, lnSD).
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}

