% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldnorm.R
\name{foldnorm}
\alias{foldnorm}
\title{foldnorm
Applying the folded normal distribution using the posterior parameter estimates of effect sizes that are normally distributed. This function will allow one to understand the overall magnitude of effect regardless of effect size direction.}
\usage{
foldnorm(mu, sd, type = c("mean", "mode"))
}
\arguments{
\item{mu}{The posterior distribution of mean estimates from an MCMCglmm object.}

\item{sd}{The standard deviation of the posterior distribution of variance-covariance matrix from an MCMCglmm object.}

\item{type}{Indicates whether the posterior mean (i.e. "mean") or mode (i.e. "mode") should be returned.}
}
\value{
Posterior mean or mode along with the 95% credible intervals (i.e. the highest posterior density interval - HPDinterval)
}
\description{
foldnorm
Applying the folded normal distribution using the posterior parameter estimates of effect sizes that are normally distributed. This function will allow one to understand the overall magnitude of effect regardless of effect size direction.
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}

