% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2.R
\name{I2}
\alias{I2}
\title{I^2 Function}
\usage{
I2(model, v, ME = FALSE, sims = 1500, phylo = FALSE)
}
\arguments{
\item{model}{The MCMCglmm or metafor model object. Note that if using a metafor model object an observation level random effect must be used to calculate the residual variance. This should be input as "~1|obs" in the random effect list.}

\item{v}{The vector of sampling variance for each effect size.}

\item{ME}{A character string with the name of the sampling error random effect. This is important if one wishes to enter the sampling variance matrix in as a sparse matrix (i.e. 'ginverse' argument) for MCMCglmm. Otherwise, assumed that the 'mev' argument is used.}

\item{sims}{The number of simulations used for calculating confidence intervals on I^2 estimates for metafor objects.}

\item{phylo}{A character string with the name of the phylogenetic random effect. Defaults to FALSE meaning that no phylogenetic heritability is calculated.}
}
\value{
A data.frame containing the relevant I^2 measures along with the 95 percent confidence / credible intervals.
}
\description{
Function for calculating commonly reported  I^2 measures from MCMCglmm model objects.  See Nakagawa and Santos (2012) for detailed explanation on the various types of I^2
}
\author{
Daniel Noble - daniel.noble@unsw.edu.au
}
\references{
Nakagawa, S. and Santos, E.S.A. (2012) Methodological issues and advances in biological meta-analysis. Evolutionary Ecology, 26:1253-1274.
}

