% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{lnSD_es}
\alias{lnSD_es}
\title{lnSD_es}
\usage{
lnSD_es(sd, n)
}
\arguments{
\item{sd}{Standard deviation of group}

\item{n}{Sample size of group}
}
\description{
Function for calculating log standard deviation effect size statistics
}
\references{
Nakagawa et al. (2015) Meta-analysis of variation: ecological and evolutionary applications and beyond. Methods in Ecology and Evolution, 6:143-152.
}
\author{
Alistair Senior - Alistair.senior@sydney.edu.au
}
