% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{lnVR_es}
\alias{lnVR_es}
\title{lnVR_es}
\usage{
lnVR_es(sd1, sd2, n1, n2)
}
\arguments{
\item{sd1}{Standard deviation of group 1}

\item{sd2}{Standard deviation of group 2}

\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}
}
\description{
Function for calculating log variation ratio effect size statistics.
}
\references{
Nakagawa et al. (2015) Meta-analysis of variation: ecological and evolutionary applications and beyond. Methods in Ecology and Evolution, 6:143-152.
}
\author{
Alistair Senior - Alistair.senior@sydney.edu.au
}
