% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_stat.R
\name{var_lnCVR}
\alias{var_lnCVR}
\title{Sampling variance for lnCVR}
\usage{
var_lnCVR(m1, m2, sd1, sd2, n1, n2, cor, Equal.E.C.Corr = TRUE)
}
\arguments{
\item{m1}{Mean of group 1}

\item{m2}{Mean of group 2}

\item{sd1}{Standard deviation of group 1}

\item{sd2}{Standard deviation of group 2}

\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}

\item{cor}{Assumed correlation between mean and variance. Used when a vector of statistics of length 1 is provided.}

\item{Equal.E.C.Corr}{Logical indicating whether the the correlation between mean and variance is equal in both group groups.}
}
\description{
Function for calculating sampling variance of log coefficient of variation ratio under different assumptions about mean-variance relationships.
}
\author{
Alistair Senior - Alistair.senior@sydney.edu.au
}
\references{
Nakagawa et al. (2015) Meta-analysis of variation: ecological and evolutionary applications and beyond. Methods in Ecology and Evolution, 6:143-152.
}

